Many of our dependencies dropped 3.9 support, resulting in uv picking
older versions for 3.9 (while retaining the latest versions for 3.10+).
Seeing as CPython themselves has dropped 3.9 support, it's probably best
to follow suit rather than use older versions for 3.9.

## CI

Python 3.9 was replaced with Python 3.14 in the CI, partially reverting
ed0a952bbc7a70bfd0c6603ae1be5cb1aa7940f1.

`rust-toolchain` used to use the since-removed commit
https://github.com/dtolnay/rust-toolchain/commit/4305c38b25d97ef35a8ad1f985ccf2d2242004f2
of the stable branch; its README [recommends] using a commit in the
master branch itself, so I have changed it to use the [1.94.1 commit of
the master branch][rust-1.94.1-commit]. I'm not sure why they recommend
that though, as GitHub doesn't garbage collect commits.

[recommends]:
https://github.com/dtolnay/rust-toolchain/blob/3c5f7ea28cd621ae0bf5283f0e981fb97b8a7af9/README.md#choice-of-full-length-commit-sha
[rust-1.94.1-commit]:
https://github.com/dtolnay/rust-toolchain/commit/3c5f7ea28cd621ae0bf5283f0e981fb97b8a7af9

<details><summary><del>I also ran Maturin to update that generated
<code>release.yml</code>…</del></summary>

~~I also ran Maturin to update that generated `release.yml`, I have **no
idea what it does** so I have no idea if the changes it made are
desirable. Lol. I've put it in a separate commit to make it easy to nuke
if it's trash. Quoting me from that commit:~~
> ~~I barely know what this does, hopefully Maturin in its infinite
wisdom didn't blunder by depending on Docker and by using tags instead
of SHAs, I guess. I'm also not asking why there's 64-bit little-endian
PowerPC CI, nor why it just added 64-bit S/390 CI 😵‍💫. It's probably
also best to not question why it sometimes uses uv, and sometimes uses
pip, I guess? And I won't even begin to bat an eye at the venv activate
script being sourced instead of `uv run` being used.~~
>
> ~~`generate-ci`'s `--pytest` flag had been deprecated; it has hence
been moved to the `pyproject.toml` as requested.~~

</details>

**Edit**: I've since reverted that, since the initially working Windows
CI decided to break itself later; I've PR'd that separately as #19.

### Updated

```perl
                    uv   v0.8.3 -> v0.11.15
      actions/checkout   v5.0.0 -> v6.0.2
dtolnay/rust-toolchain  v1.85.1 -> v1.94.1
    astral-sh/setup-uv   v6.5.0 -> v8.1.0
```

## Python

The development dependencies are no longer pinned via `pyproject.toml`,
as lockfiles exist for that purpose.

Nothing particularly interesting has changed, modulo typer.

### New

  - `annotated-doc` v0.0.4 (via `typer` addition)
  - `shellingham` v1.5.4 (via `typer` addition)
  - `rich` v15.0.0 (via `typer` addition)
  - `markdown-it-py` v4.2.0 (via `rich` addition)
  - `mdurl` v0.1.2 (via `markdown-it-py` addition)

### Replaced

- `typer-slim` v0.16.0 -> `typer` v0.25.1 (upstream dropped support; see
[here][typer-0.22.0] and [here][typer-slim-readme])

[typer-0.22.0]: https://github.com/fastapi/typer/releases/tag/0.22.0
[typer-slim-readme]:
https://github.com/fastapi/typer/tree/0.25.1/typer-slim#typer-slim

### Removed

  - `click` v8.1.8 (via Python 3.9 drop)

### Updated

```perl
            click    v8.2.1 -> v8.4.2
   exceptiongroup    v1.3.0 -> v1.3.1
        iniconfig    v2.1.0 -> v2.3.0
          nodeenv    v1.9.1 -> v1.10.0
        packaging     v25.0 -> v26.2
         pygments   v2.19.2 -> v2.20.0
          pyright  v1.1.403 -> v1.1.409
           pytest    v8.4.1 -> v9.0.3
             ruff   v0.12.9 -> v0.15.13
            tomli    v2.2.1 -> v2.4.1
typing-extensions   v4.14.1 -> v4.15.0
```

## Rust

The `from_py_object`s were added to retain compatibility with what was
present before; per the deprecation warning:
> The `FromPyObject` implementation for `#[pyclass]` types which
implement `Clone` is changing to an opt-in option. Use
> `#[pyclass(from_py_object)]` to opt-in to the `FromPyObject` derive
now, or `#[pyclass(skip_from_py_object)]` to skip the `FromPyObject`
implementation.

I'm **not sure if this is what I should do**; let me know if I should
use `skip_from_py_object` instead.

The `i as u32` seems to be the intended fix; the PR that wrought this
change, `https://github.com/tree-sitter/tree-sitter/pull/4664` (inline
code to avoid annoying cross-reference), itself does this to the values
procured from `child_count()`.

> [!WARNING]
> There were far too many updates for me to check them all; I have no
idea if anything particularly interesting changed.

### New

  - `find-msvc-tools` v0.1.9 (via `cc`; `build.rs`)
  - `serde_core` v1.0.228 (via `serde`)

### Replaced

```perl
    wasi v0.14.2+wasi-0.2.4 -> wasip2 v1.0.3+wasi-0.2.9 (via getrandom)
     wit-bindgen-rt v0.39.0 -> wit-bindgen v0.57.1      (via wasi removal)
                ryu v1.0.20 -> zmij v1.0.21             (via serde_json)
```

### Removed

  - `indoc` v2.0.6 (via `pyo3`)
  - `unindent` v0.2.4 (via `pyo3`)
  - `memoffset` v0.9.1 (via `pyo3`)
  - `autocfg` v1.5.0 (via `memoffset` removal)
  - `bitflags` v2.9.1 (via `wit-bindgen-rt` removal)

### Updated

```perl
        aho-corasick    v1.1.3 -> v1.1.4
                  cc   v1.2.30 -> v1.2.62
              cfg-if    v1.0.1 -> v1.0.4
           getrandom    v0.3.3 -> v0.3.4
           hashbrown   v0.15.4 -> v0.17.1
            indexmap   v2.10.0 -> v2.14.0
                itoa   v1.0.15 -> v1.0.18
           jobserver   v0.1.33 -> v0.1.34
                libc  v0.2.174 -> v0.2.186
                 log   v0.4.27 -> v0.4.29
              memchr    v2.7.5 -> v2.8.0
           once_cell   v1.21.3 -> v1.21.4
               pcre2    v0.2.9 -> v0.2.11
           pcre2-sys    v0.2.9 -> v0.2.10
          pkg-config   v0.3.32 -> v0.3.33
     portable-atomic   v1.11.1 -> v1.13.1
         proc-macro2   v1.0.95 -> v1.0.106
                pyo3   v0.24.2 -> v0.28.3
   pyo3-build-config   v0.24.2 -> v0.28.3
            pyo3-ffi   v0.24.2 -> v0.28.3
         pyo3-macros   v0.24.2 -> v0.28.3
 pyo3-macros-backend   v0.24.2 -> v0.28.3
       python3-dll-a   v0.2.14 -> v0.2.15
               quote   v1.0.40 -> v1.0.45
               regex   v1.11.1 -> v1.12.3
      regex-automata    v0.4.9 -> v0.4.14
        regex-syntax    v0.8.5 -> v0.8.10
               serde  v1.0.219 -> v1.0.228
        serde_derive  v1.0.219 -> v1.0.228
          serde_json  v1.0.141 -> v1.0.149
                 syn  v2.0.104 -> v2.0.117
      target-lexicon   v0.13.2 -> v0.13.5
         tree-sitter   v0.25.8 -> v0.26.9
tree-sitter-language    v0.1.5 -> v0.1.7
       unicode-ident   v1.0.18 -> v1.0.24
```

+++ src/parse.rs
@@ -86,7 +86,7 @@ fn traverse<'a>(root: tree_sitter::Node, src: &'a [u8]) -> Vec<Token<'a>> {
             stack.extend(
                 (0..node.child_count())
                     .rev()
-                    .map(|i| node.child(i).expect("should be in-bounds")),
+                    .map(|i| node.child(i as u32).expect("should be in-bounds")),
             );
             continue;
         }
